RESOLVED - task 489: (Alejandro Fdz.) Modificación tras migración a wicket 6.11 http://git.fic.udc.es/redmine/issues/show/489
Modificadas las clases que me han sido asignadas para adaptarlas a
wicket 6.11.

-- *ListDataProvider.
Cambiar parámetros del método ListDataProvider (de int a long)
Cambiar el tipo que devuelve el metodo size (de int a long)

-- En todas las demás.
Modificado el método por el cual utilizaba el CSS correcto, ya que han
cambiado los nombres de los métodos.
En las que llevan tanto miniaturas de las imágenes como las imágenes
completas se le ha añadido al metodo getBlob un parámetro, ya que ese
método sin parametros ya no existía.

+++ src/main/java/es/udc/fi/dc/photoalbum/wicket/FileListDataProvider.java
@@ -30,13 +30,13 @@ public class FileListDataProvider implements IDataProvider<File> {
 		Injector.get().inject(this);
 	}
 
-	public Iterator<File> iterator(int first, int count) {
+	public Iterator<File> iterator(long first, long count) {
 		LoadableDetachableModel<ArrayList<File>> ldm = new FilesModelPaging(
-				this.id, first, count);
+				this.id, (int) first, (int) count);
 		return ldm.getObject().iterator();
 	}
 
-	public int size() {
+	public long size() {
 		return this.size;
 	}
