+++ src/main/java/es/udc/fi/dc/photoalbum/wicket/PublicFileListDataProvider.java
@@ -31,13 +31,13 @@ public class PublicFileListDataProvider implements IDataProvider<File> {
 		Injector.get().inject(this);
 	}
 
-	public Iterator<File> iterator(int first, int count) {
+	public Iterator<File> iterator(long first, long count) {
 		LoadableDetachableModel<ArrayList<File>> ldm = new PublicFilesModelPaging(
-				this.albumId, this.userId, first, count);
+				this.albumId, this.userId, (int) first, (int) count);
 		return ldm.getObject().iterator();
 	}
 
-	public int size() {
+	public long size() {
 		return this.size;
 	}
